Basic Address Book, for coding challenge (not complete)
Unfortunately I did not have as much free time as I planned to have in order to complete this. I think it begins to show my approach when handling a wide range of areas within an app.

I would definitely perform much better and have something stronger if I were to attempt this again.